pyproject.toml
[tool.poetry]
name = "youtubeextractor"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
opencv-python = "^4.10.0.84"
yt-dlp = "^2025.1.12"
pytesseract = "^0.3.13"
openai = "^1.59.6"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
